Refactor Geometry to move vector, rotation, and pose to separate files; move other math files to a Math package
