+++ core/src/main/kotlin/gay/zharel/hermes/control/LQRController.kt
@@ -3,21 +3,21 @@
 package gay.zharel.hermes.control
 
 import gay.zharel.hermes.geometry.Acceleration2d
-import gay.zharel.hermes.geometry.DualNum
-import gay.zharel.hermes.geometry.DualParameter
-import gay.zharel.hermes.geometry.Matrix
+import gay.zharel.hermes.math.DualNum
+import gay.zharel.hermes.math.DualParameter
+import gay.zharel.hermes.math.Matrix
 import gay.zharel.hermes.geometry.Pose2d
 import gay.zharel.hermes.geometry.Pose2dDual
 import gay.zharel.hermes.geometry.PoseVelocity2d
 import gay.zharel.hermes.geometry.PoseVelocity2dDual
-import gay.zharel.hermes.geometry.Time
+import gay.zharel.hermes.math.Time
 import gay.zharel.hermes.geometry.Twist2d
 import gay.zharel.hermes.geometry.Twist2dDual
 import gay.zharel.hermes.geometry.Vector2d
 import gay.zharel.hermes.geometry.Vector2dDual
-import gay.zharel.hermes.geometry.makeBrysonMatrix
-import gay.zharel.hermes.geometry.times
-import gay.zharel.hermes.geometry.unaryMinus
+import gay.zharel.hermes.math.makeBrysonMatrix
+import gay.zharel.hermes.math.times
+import gay.zharel.hermes.math.unaryMinus
 import org.ejml.simple.SimpleMatrix
 import kotlin.time.DurationUnit
 import kotlin.time.TimeSource
