+++ core/src/main/kotlin/gay/zharel/hermes/paths/PosePaths.kt
@@ -3,6 +3,8 @@
 package gay.zharel.hermes.paths
 
 import gay.zharel.hermes.geometry.*
+import gay.zharel.hermes.math.Arclength
+import gay.zharel.hermes.math.clamp
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
