+++ core/src/main/kotlin/gay/zharel/hermes/trajectories/Markers.kt
@@ -1,6 +1,6 @@
 package gay.zharel.hermes.trajectories
 
-import gay.zharel.hermes.geometry.DualParameter
+import gay.zharel.hermes.math.DualParameter
 import gay.zharel.hermes.geometry.Pose2d
 import gay.zharel.hermes.geometry.RobotState
 import gay.zharel.hermes.geometry.Vector2d
