+++ core/src/main/kotlin/gay/zharel/hermes/trajectories/SafeBuilders.kt
@@ -2,7 +2,7 @@ package gay.zharel.hermes.trajectories
 
 import gay.zharel.hermes.profiles.AccelConstraint
 import gay.zharel.hermes.profiles.VelConstraint
-import gay.zharel.hermes.geometry.Arclength
+import gay.zharel.hermes.math.Arclength
 import gay.zharel.hermes.geometry.Pose2d
 import gay.zharel.hermes.geometry.Rotation2d
 import gay.zharel.hermes.geometry.Vector2d
