+++ core/src/main/kotlin/gay/zharel/hermes/trajectories/TrajectoryBuilder.kt
@@ -1,6 +1,6 @@
 package gay.zharel.hermes.trajectories
 
-import gay.zharel.hermes.geometry.Arclength
+import gay.zharel.hermes.math.Arclength
 import gay.zharel.hermes.geometry.Pose2d
 import gay.zharel.hermes.geometry.Rotation2d
 import gay.zharel.hermes.geometry.Vector2d
