+++ core/src/test/kotlin/gay/zharel/hermes/control/RamseteTest.kt
@@ -2,6 +2,8 @@ package gay.zharel.hermes.control
 
 import gay.zharel.hermes.TEST_PROFILE_PARAMS
 import gay.zharel.hermes.geometry.*
+import gay.zharel.hermes.math.DualParameter
+import gay.zharel.hermes.math.Time
 import gay.zharel.hermes.paths.TangentPath
 import gay.zharel.hermes.profiles.CancelableProfile
 import gay.zharel.hermes.profiles.ProfileAccelConstraint
